fix(ts): remove default export
With the way things are built, at runtime the `default` isn't set properly and isn't interpreted as a default export in any project I've tried this in.

Having this in place though means that whenever I have VSCode auto-import, it uses the default import instead of the named one and then I end up with a runtime error. So then I've got to manually switch it to a named import.

Besides, default exports are the worst.